requirements.txt
#ML
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
imageio
numpy
Pillow
scipy
opencv-python
# Code invention
git+https://github.com/psf/black
autoflake
isort
tqdm
#UI
streamlit